Add plotting_decorator and plot_or_update, bump version to 0.1.1.0 Since plotting_decorator uses doublewrap, it can be used either as:
@plotting_decorator
def someplotfunction():
    ...

or

@plotting_decorator(draw=False)
def someplotfunction():
    ...

In either case, it add options to specify a specific figure
as well as specify cla, clf, or draw options at call time:

someplotfunction(figure=myfig, draw=True, clf=True)

-------------------------------------------------------------

plot_or_update is a tool for editing data within plots dynamically
and automatically, given some existing plot result ("a" here).
When this result is None, the plot funciton is called.
When this result is not None, the data is just updated and other options are ignored

Call it like:
a = None
some loop or other context:
    a = plot_or_update(a, someplotfunction, *plot_args, **plot_kwds)

Example usage:
a = None
for i in range(5)
    a = plot_or_update(a, plt.plot, [1, 2, 3, 5, i], 'r*-')

There is an entire test suite for it in test